json problem fixed

+++ src/main/java/de/telekom/sea7/inter/repository/ReceiverRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import de.telekom.sea7.impl.repository.Receiver;
 
-public interface ReceiverRepository extends JpaRepository<Receiver, Integer> {
+public interface ReceiverRepository extends JpaRepository<Receiver, Long> {
 
 }
